requirements.txt
absl-py==2.0.0
affine==2.4.0
aiohttp==3.9.0
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
ansi2html==1.9.0rc1
anyio==4.3.0
apispec==5.2.2
apispec-webframeworks==0.5.2
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
aspose-words==24.5.0
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
binaryornot==0.4.4
bleach==6.1.0
blinker==1.7.0
branca==0.7.0
cachetools==5.3.2
category-encoders==2.6.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
constantly==23.10.4
contextily==1.6.0
contourpy==1.2.0
cookiecutter==2.1.1
cryptography==42.0.7
cycler==0.12.1
Cython==3.0.0
cytoolz==0.12.2
dacite==1.8.1
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.1.0
dataclasses-json==0.6.5
debugpy==1.8.0
decorator==4.4.2
deepdiff==6.2.3
defusedxml==0.7.1
deprecation==2.1.0
distributed==2024.1.0
distro==1.9.0
dnspython==2.4.2
ecdsa==0.19.0
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi_keycloak==1.0.11
fastjsonschema==2.18.0
filelock==3.13.1
fiona==1.9.6
Flask==3.0.1
Flask-Cors==4.0.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
folium==0.15.1
fonttools==4.45.1
frozendict==2.4.0
frozenlist==1.4.0
fsspec==2023.10.0
fundamentus==0.2.0
gast==0.4.0
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
geotiler==0.15.0
gevent==23.9.1
gevent-websocket==0.10.1
gitdb==4.0.11
gitignore_parser==0.1.11
GitPython==3.1.41
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hydrobr==0.1.1
hyperlink==21.0.0
idna==3.4
ImageHash==4.3.1
imageio==2.33.1
imageio-ffmpeg==0.4.9
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
incremental==22.10.0
ipykernel==6.26.0
ipython==8.14.0
ipywidgets==8.1.0
itsdangerous==2.1.2
jax==0.4.21
jedi==0.19.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.3.2
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kaleido==0.2.1
keras==2.12.0
kiwisolver==1.4.5
kthread==0.2.3
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langsmith==0.1.52
libclang==16.0.6
lightgbm==4.0.0
llvmlite==0.41.0rc1
locket==1.0.0
lxml==4.9.4
mailmerge==2.2.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.28.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mercantile==1.2.1
mistune==3.0.2
ml-dtypes==0.2.0
modin==0.26.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multimethod==1.11.2
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==2.8.8
numba==0.58.0rc1
numpy==1.23.5
oauthlib==3.2.2
ollama==0.1.9
openai==1.24.0
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.1
packaging==23.2
pafy==0.5.5
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
passlib==1.7.4
patsy==0.5.3
pdfkit==1.0.0
peewee==3.17.0
phik==0.12.4
pickleshare==0.7.5
Pillow==10.0.1
pivottablejs==0.9.0
platformdirs==3.10.0
plotly==5.16.1
plotly-calplot==0.1.13
plotly-resampler==0.9.1
pmdarima==2.0.3
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==10.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycaret==3.0.0rc9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydicom==2.4.4
Pygments==2.16.1
pymongo==4.6.1
pynetdicom @ git+https://github.com/pydicom/pynetdicom.git@64ead66eeb4d53c0e93fff54da5d7e8543821ecf
pyod==1.1.0
pyorthanc==1.16.1
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-jose==3.3.0
python-keycloak==4.0.0
python-multipart==0.0.9
python-slugify==8.0.1
python-socketio==5.11.0
pytz==2023.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.5.0
pywin32==306
PyYAML==6.0.1
pyzmq==24.0.1
rasterio==1.3.9
referencing==0.30.2
regex==2024.4.28
reportlab==4.2.0
requests==2.31.0
requests-cache==1.1.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.7.0
rpds-py==0.9.2
rsa==4.9
schemdraw==0.17
scikeras==0.11.0
scikit-base==0.5.1
scikit-learn==1.3.0
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.12.2
shap==0.44.0
shapely==2.0.3
shellingham==1.5.4
simple-websocket==0.10.1
six==1.16.0
sktime==0.22.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
st-star-rating==0.0.6
st-theme==1.2.3
stack-data==0.6.2
starlette==0.37.2
statsmodels==0.14.0
streamlit==1.34.0
streamlit-folium==0.18.0
streamlit-geolocation==0.0.10
streamlit-keycloak==1.1.1
streamlit-navigation-bar==3.2.0
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
taipy==3.0.0
taipy-config==3.0.0
taipy-core==3.0.0
taipy-gui==3.0.0
taipy-rest==3.0.0
taipy-templates==3.0.0
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.2
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets==5.9.0
tsdownsample==0.1.2
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typeguard==4.2.1
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==4.3.1
ujson==5.10.0
ultralytics==8.0.216
url-normalize==1.4.3
urllib3==2.1.0
uvicorn==0.29.0
validators==0.22.0
visions==0.7.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.8
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
xxhash==3.3.0
xyzservices==2023.10.1
yarl==1.9.3
ydata-profiling==4.7.0
yellowbrick==1.5
yfinance==0.2.36
youtube-dl==2020.12.2
zict==3.0.0
zipfile36==0.1.3
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
